+++ src/aks-preview/azext_aks_preview/_params.py
@@ -801,7 +801,6 @@ def load_arguments(self, _):
             "enable_advanced_network_observability",
             action="store_true",
             is_preview=True,
-            help="enable advanced network observability functionalities for cluster",
         )
         c.argument(
             "custom_ca_trust_certificates",
@@ -1296,13 +1295,11 @@ def load_arguments(self, _):
             "enable_advanced_network_observability",
             action="store_true",
             is_preview=True,
-            help="enable advanced network observability functionalities for cluster",
         )
         c.argument(
             "disable_advanced_network_observability",
             action="store_true",
             is_preview=True,
-            help="disable advanced network observability functionalities for cluster",
         )
         c.argument("enable_cost_analysis", is_preview=True, action="store_true")
         c.argument("disable_cost_analysis", is_preview=True, action="store_true")